Finish screenshots on test failure
This method uses defmacro to override test with our own testcase
function, sstest. Google was helpful to get me started but I had to
fix a lot of problems.

Last part of this I did not complete was generating the folder based on
the test name for each screenshot, and naming the image after the
current date time UTC. All of my attempts to do so failed. Remember to
ask interviewers how they would do this.